getStatsList API

Search for statistical datasets via e-Stat API.

estat_getStatsList(
  appId,
  searchWord,
  lang = c("J", "E"),
  .use_label = TRUE,
  surveyYears = NULL,
  openYears = NULL,
  statsField = NULL,
  statsCode = NULL,
  searchKind = NULL,
  startPosition = NULL,
  limit = NULL,
  updatedDate = NULL,
  ...
)

Arguments

appId
Application ID.

searchWord
Keyword for searching. You can use OR and AND to combine conditions (e.g.: "apple AND orange").

lang
Language of the data. "J"(Japanese) or "E"(English).

.use_label
Whether to take the human-redable label value or the code value when flattening a field containing both (default: TRUE).

surveyYears
Year and month when the survey was conducted. The format is either "YYYY", "YYYYMM", or "YYYYMM-YYYYMM".

openYears
Year and month when the survey result was opened. The format is the same as surveyYears.

statsField
Field of statistics. The format is either two digits(large classification) or four digits (small classification). For the detail of the classification, see http://www.soumu.go.jp/toukei_toukatsu/index/seido/sangyo/26index.htm.

statsCode
Code assigned for each statistical agency and statistics. The format can be five digits (agency), and eight digits (statistics). For the detail, see http://www.stat.go.jp/info/guide/public/code/code.htm.

searchKind
Type of statistics. 1: summary, 2: regional mesh.

startPosition
Index of the first record to get.

limit
Max number of records to get.

updatedDate
Last updated date. The format is either "YYYY", "YYYYMM", or "YYYYMM-YYYYMM".

...
Other parameters.

See also

Examples

if (FALSE) {
estat_getStatsList(
  appId = "XXXX",
  searchWord = "CD",
  limit = 3
)
}